requirements.txt
fastapi
uvicorn
pillow
requests
websockets
